serving_api/requirements.txt
flask
flask-cors
python-dotenv
mlflow
onnxruntime